Feature addprogressbar (#19)
* Added file headers on all files

* Bug Fix - Fixed crash on empty URL (#18)

Fixed a crash when user attempted to load an empty Url. The log will now reflect the error and tell the user that the Url is empty

* Added IsLoading property to track loading progress

This property is set to true at the beginning of an operation and false at the end of the operation. Controls the visibility (Turn on/off) of the progress bar

* Implemented IsLoading property to operations

Wrapped State.IsLoading property around operations

* Added a progress bar across the UI above the log

I tried to maintain the same structure of prior UI code. I added a row above the Log textblock and shifted down the Textbox and Listbox by 1 index in the row. The data that controls the visibility of the progress bar was bound to State.IsLoading property

Co-authored-by: Chad Birch <chadbirch@hotmail.com>